requirements.txt
python-telegram-bot
mistralai
httpx>=0.27.0
python-dotenv==1.1.0 